+++ plugins/source/aws/policies/queries/iam/password_policy_expire_old_passwords.sql
@@ -1,8 +1,8 @@
 insert into aws_policy_results
 select
-  :'execution_time',
-  :'framework',
-  :'check_id',
+  :'execution_time' as execution_time,
+  :'framework' as framework,
+  :'check_id' as check_id,
   'Ensure IAM password policy expires passwords within 90 days or less' as title,
   account_id,
   account_id,