+++ plugins/source/aws/policies/queries/iam/password_policy_min_length.sql
@@ -1,9 +1,9 @@
 insert into aws_policy_results
 select
-  :'execution_time',
-  :'framework',
-  :'check_id',
-  'Ensure IAM password policy requires minimum length of 14 or greater',
+  :'execution_time' as execution_time,
+  :'framework' as framework,
+  :'check_id' as check_id,
+  'Ensure IAM password policy requires minimum length of 14 or greater' as title,
   account_id,
   account_id,
   case when