+++ plugins/source/aws/policies/queries/iam/wildcard_access_policies.sql
@@ -34,10 +34,10 @@ allow_all_statements as (
 )
 
 select distinct
-    :'execution_time'::timestamp,
-    :'framework',
-    :'check_id',
-    'IAM customer managed policies that you create should not allow wildcard actions for services' AS title,
+    :'execution_time'::timestamp as execution_time,
+    :'framework' as framework,
+    :'check_id' as check_id,
+    'IAM customer managed policies that you create should not allow wildcard actions for services' as title,
     aws_iam_policies.account_id,
     aws_iam_policies.arn AS resource_id,
     CASE WHEN statements_count > 0 THEN 'fail' ELSE 'pass' END AS status