+++ plugins/source/azure/policies/queries/sql/mysql_ssl_enforcment_disabled.sql
@@ -1,9 +1,9 @@
 insert into azure_policy_results
 SELECT
-  :'execution_time',
-  :'framework',
-  :'check_id',
-  'Ensure "Enforce SSL connection" is set to "ENABLED" for MySQL Database Server (Automated)',
+  :'execution_time' as execution_time,
+  :'framework' as framework,
+  :'check_id' as check_id,
+  'Ensure "Enforce SSL connection" is set to "ENABLED" for MySQL Database Server (Automated)' as title,
   subscription_id,
   id AS server_id,
   case