+++ plugins/source/azure/policies/queries/sql/va_periodic_scans_enabled_on_sql_server.sql
@@ -1,9 +1,9 @@
 insert into azure_policy_results
 SELECT 
-  :'execution_time',
-  :'framework',
-  :'check_id',
-  'Ensure that VA setting Periodic Recurring Scans is enabled on a SQL server (Automated)',
+  :'execution_time' as execution_time,
+  :'framework' as framework,
+  :'check_id' as check_id,
+  'Ensure that VA setting Periodic Recurring Scans is enabled on a SQL server (Automated)' as title,
   s.subscription_id,
   s.id,
   case