+++ plugins/source/gcp/resources/services/compute/projects.go
@@ -3,13 +3,8 @@
 package compute
 
 import (
-	"context"
-	"github.com/pkg/errors"
-
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/cloudquery/plugins/source/gcp/client"
-
-	pb "google.golang.org/genproto/googleapis/cloud/compute/v1"
 )
 
 func Projects() *schema.Table {
@@ -92,17 +87,4 @@ func Projects() *schema.Table {
 			},
 		},
 	}
-}
-
-func fetchProjects(ctx context.Context, meta schema.ClientMeta, parent *schema.Resource, res chan<- interface{}) error {
-	c := meta.(*client.Client)
-	req := &pb.GetProjectRequest{
-		Project: c.ProjectId,
-	}
-	resp, err := c.Services.ComputeProjectsClient.Get(ctx, req)
-	if err != nil {
-		return errors.WithStack(err)
-	}
-	res <- resp
-	return nil
 }