+++ plugins/source/gcp/resources/services/compute/vpn_gateways_mock_test.go
@@ -3,31 +3,24 @@
 package compute
 
 import (
-	"context"
 	"encoding/json"
 	"net/http"
-	"net/http/httptest"
 	"testing"
 
 	"github.com/cloudquery/plugin-sdk/faker"
 	"github.com/cloudquery/plugins/source/gcp/client"
 	"github.com/julienschmidt/httprouter"
 
-	"cloud.google.com/go/compute/apiv1"
-
 	pb "google.golang.org/genproto/googleapis/cloud/compute/v1"
-
-	"google.golang.org/api/option"
 )
 
-func createVpnGateways() (*client.Services, error) {
+func createVpnGateways(mux *httprouter.Router) error {
 	var item pb.VpnGatewayAggregatedList
 	if err := faker.FakeObject(&item); err != nil {
-		return nil, err
+		return err
 	}
 	emptyStr := ""
 	item.NextPageToken = &emptyStr
-	mux := httprouter.New()
 	mux.GET("/*filepath", func(w http.ResponseWriter, r *http.Request, _ httprouter.Params) {
 		b, err := json.Marshal(&item)
 		if err != nil {
@@ -39,16 +32,9 @@ func createVpnGateways() (*client.Services, error) {
 			return
 		}
 	})
-	ts := httptest.NewServer(mux)
-	svc, err := compute.NewVpnGatewaysRESTClient(context.Background(), option.WithoutAuthentication(), option.WithEndpoint(ts.URL))
-	if err != nil {
-		return nil, err
-	}
-	return &client.Services{
-		ComputeVpnGatewaysClient: svc,
-	}, nil
+	return nil
 }
 
 func TestVpnGateways(t *testing.T) {
-	client.MockTestHelper(t, VpnGateways(), createVpnGateways, client.TestOptions{})
+	client.MockTestRestHelper(t, VpnGateways(), createVpnGateways, client.TestOptions{})
 }